+++ RELEASE-NOTES.md
@@ -6,6 +6,7 @@
 * #608 selectAll(Udf1)
 * #609 ECharts: minify generated JavaScript
 * #612 "concat" / "vConcat" naming ambiguity
+* #613 Recast hConcat(..) as positional join
 
 ## Release 2.0.0-M6